Update in project's structure and using email as pk in database

+++ src/controller/user/userController.js
@@ -1,33 +1,40 @@
-import jwt from 'jsonwebtoken';
-import admin from '../../database/connection';
-
-require('dotenv').config();
+import admin from '../../configs/database/connection';
 
 const db = admin.firestore();
+
+async function getUsuario(email) {
+  const userCollection = db.collection('user');
+  let dbVerification = null;
+  await userCollection
+    .where('email', '==', email)
+    .get()
+    .then((snapshot) => {
+      return snapshot.forEach((res) => {
+        dbVerification = res.id;
+      });
+    });
+  if (!dbVerification) {
+    return null;
+  }
+  return dbVerification;
+}
 
 module.exports = {
   async get(request, response) {
     try {
-      const { user } = request.body;
-      if (!user) {
+      const userCollection = db.collection('user');
+      const results = [];
+      await userCollection.get().then((snapshot) => {
+        snapshot.forEach((doc) => {
+          results.push(doc.data());
+        });
+      });
+      if (!results.length) {
         return response
           .status(400)
-          .json({ error: 'Não foram enviados os dados.' });
+          .json({ error: 'Não tem usuários para serem listados' });
       }
-      const userCollection = db.collection('user');
-      let result = null;
-      await userCollection
-        .where('usuario', '==', user)
-        .get()
-        .then((snapshot) => {
-          return snapshot.forEach((res) => {
-            result = res.data();
-          });
-        });
-      if (!result) {
-        return response.status(400).json({ error: 'Não foi encontrado.' });
-      }
-      return response.status(200).json(result);
+      return response.status(200).json(results);
     } catch (e) {
       return response.status(500).json({
         error: `Erro durante o processamento do login. Espere um momento e tente novamente! Erro : ${e}`,
@@ -37,7 +44,6 @@ module.exports = {
   async insert(request, response) {
     try {
       const {
-        user,
         cpf,
         mentorFlag,
         email,
@@ -49,21 +55,12 @@ module.exports = {
 
       const userCollection = db.collection('user');
 
-      let dbVerification = null;
-      await userCollection
-        .where('user', '==', user)
-        .get()
-        .then((snapshot) => {
-          return snapshot.forEach((res) => {
-            dbVerification = res.data();
-          });
-        });
+      const dbVerification = await getUsuario(email);
       if (dbVerification) {
         return response.status(400).send({ error: 'Usuário já existe.' });
       }
 
       await userCollection.add({
-        user,
         password,
         name,
         cpf,
@@ -72,7 +69,7 @@ module.exports = {
         email,
         mentorFlag,
       });
-      return response.status(201).send();
+      return response.status(200).send({ success: true });
     } catch (e) {
       return response.status(500).json({
         error: `Erro durante o processamento do login. Espere um momento e tente novamente! Erro : ${e}`,
@@ -83,7 +80,6 @@ module.exports = {
   async update(request, response) {
     try {
       const {
-        user,
         cpf,
         mentorFlag,
         email,
@@ -94,22 +90,13 @@ module.exports = {
       } = request.body;
 
       const userCollection = db.collection('user');
+      const dbVerification = await getUsuario(email);
 
-      let dbVerification = null;
-      await userCollection
-        .where('user', '==', user)
-        .get()
-        .then((snapshot) => {
-          return snapshot.forEach((res) => {
-            dbVerification = res.id;
-          });
-        });
       if (!dbVerification) {
         return response.status(400).send({ error: 'Usuário não existe.' });
       }
 
-      await userCollection.doc(dbVerification).set({
-        user,
+      const teste = await userCollection.doc(dbVerification).set({
         password,
         name,
         cpf,
@@ -118,6 +105,8 @@ module.exports = {
         email,
         mentorFlag,
       });
+
+      console.log(teste);
       return response.status(200).send();
     } catch (e) {
       return response.status(500).json({
@@ -128,68 +117,24 @@ module.exports = {
 
   async delete(request, response) {
     try {
-      const { user } = request.body;
+      const { email } = request.body;
+
+      if (!email) {
+        return response
+          .status(400)
+          .send({ error: 'Variável email dever ser passada ' });
+      }
 
       const userCollection = db.collection('user');
-
-      let dbVerification = null;
-      await userCollection
-        .where('user', '==', user)
-        .get()
-        .then((snapshot) => {
-          return snapshot.forEach((res) => {
-            dbVerification = res.id;
-          });
-        });
+      const dbVerification = await getUsuario(email);
       if (!dbVerification) {
         return response.status(400).send({ error: 'Usuário não existe.' });
       }
 
       await userCollection.doc(dbVerification).delete();
-      return response.status(200).send();
-    } catch (e) {
-      return response.status(500).json({
-        error: `Erro durante o processamento do login. Espere um momento e tente novamente! Erro : ${e}`,
-      });
-    }
-  },
-
-  async login(request, response) {
-    try {
-      const { user, password } = request.body;
-
-      if (!user || !password) {
-        return response
-          .status(404)
-          .json({ error: 'Não foram enviados os dados.' });
-      }
-      const userCollection = db.collection('user');
-      let result = null;
-      await userCollection
-        .where('usuario', '==', user)
-        .where('password', '==', password)
-        .get()
-        .then((snapshot) => {
-          return snapshot.forEach((res) => {
-            result = res.data();
-          });
-        });
-
-      if (!result) {
-        return response.status(401).json({ error: 'Erro de autenticação!' });
-      }
-      const token = jwt.sign(
-        {
-          cpf: result.cpf,
-          email: result.email,
-        },
-        process.env.JWT_KEY,
-        {
-          expiresIn: '1h',
-        }
-      );
-
-      return response.status(200).json({ token });
+      return response
+        .status(200)
+        .send({ success: true, msg: `${email} removido com sucesso!` });
     } catch (e) {
       return response.status(500).json({
         error: `Erro durante o processamento do login. Espere um momento e tente novamente! Erro : ${e}`,
